# packages/api-contract/generated/openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sanskrit Shloka Learning API",
    "version": "0.0.0"
  },
...
  "paths": {
...
    "/api/dashboard": {
      "get": {
        "operationId": "Dashboard_getDashboard",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanskritShlokaLearning.EmptyDashboardDto"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanskritShlokaLearning.ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/api/dashboard/learning-shlokas": {
      "get": {
        "operationId": "Dashboard_getLearningShlokas",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "explode": false
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanskritShlokaLearning.DashboardLearningShlokaListDto"
                }
              }
            }
          },
          "400": {
            "description": "The server could not understand the request due to invalid syntax.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanskritShlokaLearning.ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanskritShlokaLearning.ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/api/dashboard/review-shlokas": {
      "get": {
        "operationId": "Dashboard_getReviewShlokas",
        "parameters": [
          {
            "name": "timeZone",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "explode": false
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "explode": false
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanskritShlokaLearning.DashboardReviewShlokaListDto"
                }
              }
            }
          },
          "400": {
            "description": "The server could not understand the request due to invalid syntax.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanskritShlokaLearning.ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanskritShlokaLearning.ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/api/dashboard/streak": {
      "get": {
        "operationId": "Dashboard_getStreak",
        "parameters": [
          {
            "name": "timeZone",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "explode": false
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanskritShlokaLearning.DashboardStreakDto"
                }
              }
            }
          },
          "400": {
            "description": "The server could not understand the request due to invalid syntax.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanskritShlokaLearning.ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanskritShlokaLearning.ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/api/library": {
      "get": {
        "operationId": "Library_getLibrary",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanskritShlokaLearning.LibraryResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanskritShlokaLearning.ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "Library"
        ]
      }
    },
    "/api/library/items/{shlokaCode}": {
      "get": {
        "operationId": "Library_getItem",
        "parameters": [
          {
            "name": "shlokaCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanskritShlokaLearning.LibraryShlokaDto"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanskritShlokaLearning.ApiError"
                }
              }
            }
          },
          "404": {
            "description": "The server cannot find the requested resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanskritShlokaLearning.ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "Library"
        ]
      },
      "patch": {
        "operationId": "Library_updateItem",
        "parameters": [
          {
            "name": "shlokaCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanskritShlokaLearning.LibraryShlokaDto"
                }
              }
            }
          },
          "400": {
            "description": "The server could not understand the request due to invalid syntax.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanskritShlokaLearning.ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanskritShlokaLearning.ApiError"
                }
              }
            }
          },
          "404": {
            "description": "The server cannot find the requested resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanskritShlokaLearning.ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "Library"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SanskritShlokaLearning.UpdateLibraryItemRequest"
              }
            }
          }
        }
      }
    },
    "/api/library/items/{shlokaCode}/complete-learning": {
      "post": {
        "operationId": "Library_completeLearning",
        "parameters": [
          {
            "name": "shlokaCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanskritShlokaLearning.CompleteLearningDto"
                }
              }
            }
          },
          "400": {
            "description": "The server could not understand the request due to invalid syntax.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanskritShlokaLearning.ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanskritShlokaLearning.ApiError"
                }
              }
            }
          },
          "404": {
            "description": "The server cannot find the requested resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanskritShlokaLearning.ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "Library"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SanskritShlokaLearning.CompleteLearningRequestBody"
              }
            }
          }
        }
      }
    },
    "/api/library/items/{shlokaCode}/complete-review": {
      "post": {
        "operationId": "Library_completeReview",
        "parameters": [
          {
            "name": "shlokaCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "The request has succeeded and a new resource has been created as a result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanskritShlokaLearning.CompletedReviewDto"
                }
              }
            }
          },
          "400": {
            "description": "The server could not understand the request due to invalid syntax.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanskritShlokaLearning.ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanskritShlokaLearning.ApiError"
                }
              }
            }
          },
          "404": {
            "description": "The server cannot find the requested resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanskritShlokaLearning.ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "Library"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SanskritShlokaLearning.CompleteReviewRequest"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "SanskritShlokaLearning.ApiError": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/SanskritShlokaLearning.ErrorCode"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "SanskritShlokaLearning.CompleteLearningDto": {
        "type": "object",
        "required": [
          "shloka",
          "remainingLearningShlokas"
        ],
        "properties": {
          "shloka": {
            "$ref": "#/components/schemas/SanskritShlokaLearning.LibraryShlokaDto"
          },
          "remainingLearningShlokas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SanskritShlokaLearning.LibraryShlokaDto"
            }
          }
        }
      },
      "SanskritShlokaLearning.CompleteLearningRequestBody": {
        "type": "object",
        "required": [
          "timeZone"
        ],
        "properties": {
          "timeZone": {
            "type": "string"
          }
        }
      },
      "SanskritShlokaLearning.CompleteReviewRequest": {
        "type": "object",
        "required": [
          "result",
          "timeZone"
        ],
        "properties": {
          "result": {
            "$ref": "#/components/schemas/SanskritShlokaLearning.ReviewResult"
          },
          "timeZone": {
            "type": "string"
          }
        }
      },
      "SanskritShlokaLearning.CompletedReviewDto": {
        "type": "object",
        "required": [
          "shlokaCode",
          "completedAt",
          "userDay",
          "result"
        ],
        "properties": {
          "shlokaCode": {
            "type": "string"
          },
          "completedAt": {
            "type": "string"
          },
          "userDay": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/SanskritShlokaLearning.ReviewResult"
          }
        }
      },
...
      "SanskritShlokaLearning.DashboardLearningShlokaListDto": {
        "type": "object",
        "required": [
          "items",
          "remainingCount",
          "hasLearningShlokas"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SanskritShlokaLearning.DashboardShlokaDto"
            }
          },
          "remainingCount": {
            "type": "integer",
            "format": "int32"
          },
          "hasLearningShlokas": {
            "type": "boolean"
          }
        }
      },
      "SanskritShlokaLearning.DashboardPrimaryActionDto": {
        "type": "object",
        "required": [
          "label",
          "target"
        ],
        "properties": {
          "label": {
            "type": "string",
            "enum": [
              "Добавить"
            ]
          },
          "target": {
            "type": "string",
            "enum": [
              "/library"
            ]
          }
        }
      },
      "SanskritShlokaLearning.DashboardReviewListState": {
        "type": "string",
        "enum": [
          "empty",
          "active",
          "completed"
        ]
      },
      "SanskritShlokaLearning.DashboardReviewShlokaListDto": {
        "type": "object",
        "required": [
          "items",
          "remainingCount",
          "hasReviewingShlokas",
          "state"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SanskritShlokaLearning.DashboardShlokaDto"
            }
          },
          "remainingCount": {
            "type": "integer",
            "format": "int32"
          },
          "hasReviewingShlokas": {
            "type": "boolean"
          },
          "state": {
            "$ref": "#/components/schemas/SanskritShlokaLearning.DashboardReviewListState"
          }
        }
      },
      "SanskritShlokaLearning.DashboardShlokaDto": {
        "type": "object",
        "required": [
          "code",
          "displayTitle",
          "text"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "displayTitle": {
            "type": "string"
          },
          "text": {
            "type": "string"
          }
        }
      },
      "SanskritShlokaLearning.DashboardStreakDto": {
        "type": "object",
        "required": [
          "days",
          "continuedToday"
        ],
        "properties": {
          "days": {
            "type": "integer",
            "format": "int32"
          },
          "continuedToday": {
            "type": "boolean"
          }
        }
      },
      "SanskritShlokaLearning.EmptyDashboardDto": {
        "type": "object",
        "required": [
          "hasPersonalShlokas",
          "showStreak",
          "showReviewBlock",
          "primaryAction"
        ],
        "properties": {
          "hasPersonalShlokas": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "showStreak": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "showReviewBlock": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "primaryAction": {
            "$ref": "#/components/schemas/SanskritShlokaLearning.DashboardPrimaryActionDto"
          }
        }
      },
      "SanskritShlokaLearning.ErrorCode": {
        "type": "string",
        "enum": [
          "VALIDATION_ERROR",
          "INVALID_CREDENTIALS",
          "EMAIL_ALREADY_REGISTERED",
          "UNAUTHORIZED",
          "FORBIDDEN",
          "NOT_FOUND"
        ]
      },
      "SanskritShlokaLearning.LibraryResponseDto": {
        "type": "object",
        "required": [
          "defaultTab",
          "tabs",
          "allShlokas"
        ],
        "properties": {
          "defaultTab": {
            "$ref": "#/components/schemas/SanskritShlokaLearning.LibraryTab"
          },
          "tabs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SanskritShlokaLearning.LibraryTabDto"
            }
          },
          "allShlokas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SanskritShlokaLearning.LibraryShlokaDto"
            }
          }
        }
      },
      "SanskritShlokaLearning.LibraryShlokaDto": {
        "type": "object",
        "required": [
          "code",
          "displayTitle",
          "sourceTitle",
          "number",
          "text",
          "personalStatus"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "displayTitle": {
            "type": "string"
          },
          "sourceTitle": {
            "type": "string"
          },
          "number": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "personalStatus": {
            "$ref": "#/components/schemas/SanskritShlokaLearning.LibraryShlokaStatus"
          },
          "fullTranslation": {
            "type": "string"
          }
        }
      },
      "SanskritShlokaLearning.LibraryShlokaStatus": {
        "type": "string",
        "enum": [
          "available",
          "learning",
          "reviewing"
        ]
      },
      "SanskritShlokaLearning.LibraryTab": {
        "type": "string",
        "enum": [
          "reviewing",
          "learning",
          "all"
        ]
      },
      "SanskritShlokaLearning.LibraryTabDto": {
        "type": "object",
        "required": [
          "id",
          "label",
          "emptyTitle",
          "emptyDescription"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/SanskritShlokaLearning.LibraryTab"
          },
          "label": {
            "type": "string"
          },
          "emptyTitle": {
            "type": "string"
          },
          "emptyDescription": {
            "type": "string"
          }
        }
      },
...
      "SanskritShlokaLearning.ReviewResult": {
        "type": "string",
        "enum": [
          "remembered_without_error",
          "remembered_with_error",
          "remembered_with_hint",
          "forgot"
        ]
      },
...
      "SanskritShlokaLearning.UpdateLibraryItemRequest": {
        "type": "object",
        "required": [
          "personalStatus"
        ],
        "properties": {
          "personalStatus": {
            "$ref": "#/components/schemas/SanskritShlokaLearning.UpdateLibraryShlokaStatus"
          }
        }
      },
      "SanskritShlokaLearning.UpdateLibraryShlokaStatus": {
        "type": "string",
        "enum": [
          "available",
          "learning"
        ]
      },
...
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development API",
      "variables": {}
    }
  ]
}